feat: Update ElementComponents styles for better layout
This commit updates the styles of the ElementComponents in the code changes to improve the layout. The changes include adding the "m-10" class to the "mt-30" div in the "elementDialog" and "elementMessage" components, and adding a new "w-100vw h-2 bg-#E5E7EC my-15" div in the "index" component. These style updates enhance the visual presentation and spacing of the ElementComponents.