feat: Add vue-json-pretty package for JSON editing
This commit adds the vue-json-pretty package to the project as a new npm dependency. The vue-json-pretty package version 2.4.0 has been added to the package.json and package-lock.json files. This enhancement improves the application by providing a JSON editor component, based on vue-json-pretty, which allows users to edit and view JSON data easily.