feat: Add support for QR code generation
This commit adds support for generating QR codes using the qrcode package. The language files (en.ts and zh.ts) have been updated to include translations for the QR code component. The router configuration has been updated to add the QR code route and link it to the corresponding component. This enhancement improves the application by providing the ability to generate and display QR codes based on qrcode.