feat: Add bullet screen components and translations
This submission adds a new bullet screen component for playing the video in the application. The language files (en.ts and zh.ts) have been updated to include translations of the bullet screen components. The routing configuration has been updated to add the bullet-screen route and link it to the appropriate component. This improvement improves the application by providing localized text for playing videos and displaying the bullet screen components.